+++ packages/middleware-multipart/src/index.ts
@@ -1,3 +1,2 @@
 export * from './multipart.middleware';
 export * from './multipart.interface';
-export { StreamHandler, ParserOpts } from './multipart.parser';
