+++ packages/middleware-multipart/src/multipart.middleware.ts
@@ -1,9 +1,9 @@
 import { HttpRequest } from '@marblejs/core';
 import { of, Observable } from 'rxjs';
 import { mergeMap, map } from 'rxjs/operators';
-import { parseMultipart, ParserOpts } from './multipart.parser';
+import { parseMultipart } from './multipart.parser';
 import { shouldParseMultipart } from './multipart.util';
-import { WithFile } from './multipart.interface';
+import { WithFile, ParserOpts } from './multipart.interface';
 
 export const multipart$ = <File extends string>(opts: ParserOpts = {}) => <T extends HttpRequest>(req$: Observable<T>) =>
   req$.pipe(
