+++ packages/middleware-multipart/src/multipart.parser.ts
@@ -1,112 +1,10 @@
-import { HttpRequest, HttpError, HttpStatus } from '@marblejs/core';
-import { isNonNullable } from '@marblejs/core/dist/+internal/utils';
-import { fromReadableStream } from '@marblejs/core/dist/+internal/observable';
-import { fromEvent, Observable, of, throwError, merge, iif } from 'rxjs';
-import { mapTo, map, mergeMap, takeUntil, toArray, tap, catchError, buffer, mergeMapTo, take, ignoreElements } from 'rxjs/operators';
 import * as Busboy from 'busboy';
-import { Readable } from 'stream';
-
-type FileEvent = [string, NodeJS.ReadableStream, string, string, string];
-
-type FieldEvent = [string, any, boolean, boolean, string, string];
-
-type FileData = {
-  file: NodeJS.ReadableStream;
-  filename: string;
-  fieldname: string;
-  encoding: string;
-  mimetype: string;
-};
-
-interface StreamHandlerOutput {
-  destination: any;
-  size?: number;
-}
-
-export interface StreamHandler {
-  (opts: FileData):
-    | Promise<StreamHandlerOutput>
-    | Observable<StreamHandlerOutput>;
-}
-
-export interface ParserOpts {
-  files?: string[];
-  stream?: StreamHandler;
-  maxFileSize?: number;
-  maxFileCount?: number;
-}
-
-const fileSizeLimit = (data: FileData, maxBytes: number | undefined) =>
-  fromEvent(data.file, 'limit').pipe(
-    take(1),
-    mergeMapTo(throwError(
-      new HttpError(`Reached file size limit for "${data.fieldname}" [${maxBytes} bytes]`, HttpStatus.PRECONDITION_FAILED),
-    )),
-    ignoreElements(),
-  );
-
-const parseFile = (req: HttpRequest) => (opts: ParserOpts) => (event$: Observable<FileEvent>, finish$: Observable<any>) =>
-  event$.pipe(
-    takeUntil(finish$),
-    map(([ fieldname, file, filename, encoding, mimetype ]) => ({ fieldname, file, filename, encoding, mimetype })),
-    mergeMap(data => iif(
-      () => !!opts.files ? !opts.files.includes(data.fieldname) : false,
-      throwError(new HttpError(`File "${data.fieldname}" is not acceptable`, HttpStatus.PRECONDITION_FAILED)),
-      of(data),
-    )),
-    mergeMap(data => merge(fileSizeLimit(data, opts.maxFileSize), of(data))),
-    mergeMap(data => isNonNullable(opts.stream)
-      ? of(data).pipe(
-          mergeMap(opts.stream),
-          catchError(error => throwError(new HttpError(error.message, HttpStatus.INTERNAL_SERVER_ERROR))),
-          tap(({ destination, size }) => {
-            req.file = {
-              ...req.file || {},
-              [data.fieldname]: {
-                size,
-                destination,
-                encoding: data.encoding,
-                mimetype: data.mimetype,
-                filename: data.filename,
-                fieldname: data.fieldname,
-              },
-            };
-          }),
-          mapTo(data),
-        )
-      : of(data.file as Readable).pipe(
-          mergeMap(fromReadableStream),
-          toArray(),
-          map(chunks => Buffer.concat(chunks)),
-          catchError(error => throwError(new HttpError(error.message, HttpStatus.INTERNAL_SERVER_ERROR))),
-          tap(buffer => {
-            req.file = {
-              ...req.file || {},
-              [data.fieldname]: {
-                size: Buffer.byteLength(buffer),
-                buffer,
-                encoding: data.encoding,
-                mimetype: data.mimetype,
-                filename: data.filename,
-                fieldname: data.fieldname,
-              },
-            };
-          }),
-          mapTo(data),
-        ),
-    ),
-  );
-
-const parseField = (req: HttpRequest) => (event$: Observable<FieldEvent>, finish$: Observable<any>) =>
-  event$.pipe(
-    takeUntil(finish$),
-    tap(([ fieldname, value ]) => {
-      req.body = {
-        ...req.body || {},
-        [fieldname]: value,
-      };
-    }),
-  );
+import { HttpRequest, HttpError, HttpStatus } from '@marblejs/core';
+import { fromEvent, Observable, of, throwError, merge } from 'rxjs';
+import { mapTo, mergeMap, tap, buffer, mergeMapTo } from 'rxjs/operators';
+import { ParserOpts } from './multipart.interface';
+import { parseField } from './multipart.parser.field';
+import { parseFile } from './multipart.parser.file';
 
 export const parseMultipart = (opts: ParserOpts = {}) => (req: HttpRequest): Observable<HttpRequest> => {
   const busboy = new Busboy({
@@ -114,6 +12,8 @@ export const parseMultipart = (opts: ParserOpts = {}) => (req: HttpRequest): Obs
     limits: {
       fileSize: opts.maxFileSize,
       files: opts.maxFileCount,
+      fields: opts.maxFieldCount,
+      fieldSize: opts.maxFieldSize,
     },
   });
 
@@ -133,9 +33,20 @@ export const parseMultipart = (opts: ParserOpts = {}) => (req: HttpRequest): Obs
     )),
   );
 
+  const fieldsLimit$ = fromEvent(busboy, 'fieldsLimit').pipe(
+    mergeMap(() => throwError(
+      new HttpError(`Reached max fields count limit [${opts.maxFieldCount}]`, HttpStatus.PRECONDITION_FAILED),
+    )),
+  );
+
   return of(req).pipe(
     tap(() => req.pipe(busboy)),
-    mergeMapTo(merge(parseFile$, parseField$, filesLimit$)),
+    mergeMapTo(merge(
+      parseFile$,
+      parseField$,
+      filesLimit$,
+      fieldsLimit$,
+    )),
     buffer(fromEvent(busboy, 'finish')),
     mapTo(req),
   );
