+++ packages/middleware-multipart/src/multipart.util.ts
@@ -1,11 +1,35 @@
 import { HttpRequest } from '@marblejs/core';
 import { getContentType } from '@marblejs/core/dist/+internal/http';
+import { FileIncomingData } from './multipart.interface';
 
-const isProperMethod = (req: HttpRequest): boolean =>
+type ComputedFileData = Partial<{ buffer: Buffer; destination: any; size: number }>;
+
+export const isProperMethod = (req: HttpRequest): boolean =>
   ['POST', 'PUT'].includes(req.method);
 
-const isMultipart = (req: HttpRequest): boolean =>
+export const isMultipart = (req: HttpRequest): boolean =>
   getContentType(req.headers).includes('multipart/');
+
+export const isFieldnameParseable = (files: string[] | undefined) => (fieldname: string) =>
+  !!files ? files.includes(fieldname) : true;
 
 export const shouldParseMultipart = (req: HttpRequest) =>
   isProperMethod(req) && isMultipart(req);
+
+export const setRequestData =
+  (req: HttpRequest) =>
+  (incomingFile: FileIncomingData) =>
+  ({ buffer, destination, size }: ComputedFileData) => {
+    req.file = {
+      ...req.file || {},
+      [incomingFile.fieldname]: {
+        size,
+        buffer,
+        destination,
+        encoding: incomingFile.encoding,
+        mimetype: incomingFile.mimetype,
+        filename: incomingFile.filename,
+        fieldname: incomingFile.fieldname,
+      },
+    };
+}
